requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
transformers
accelerate
torch
pymupdf
python-docx
numpy
